moves test_common to testutils
