+++ x/qgb/keeper/keeper_valset_test.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"testing"
 
-	"github.com/celestiaorg/celestia-app/x/qgb/keeper"
+	"github.com/celestiaorg/celestia-app/testutil"
 	"github.com/celestiaorg/celestia-app/x/qgb/types"
 	gethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
@@ -54,7 +54,7 @@ func TestCurrentValsetNormalization(t *testing.T) {
 	for msg, spec := range specs {
 		spec := spec
 		t.Run(msg, func(t *testing.T) {
-			input, ctx := keeper.SetupTestChain(t, spec.srcPowers)
+			input, ctx := testutil.SetupTestChain(t, spec.srcPowers)
 			r, err := input.QgbKeeper.GetCurrentValset(ctx)
 			require.NoError(t, err)
 			rMembers, err := types.BridgeValidators(r.Members).ToInternal()
