+++ x/qgb/keeper/query_valset_confirm_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/celestiaorg/celestia-app/x/qgb/keeper"
+	"github.com/celestiaorg/celestia-app/testutil"
 	"github.com/celestiaorg/celestia-app/x/qgb/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -23,7 +23,7 @@ func TestQueryValsetConfirm(t *testing.T) {
 	)
 	require.NoError(t, err1)
 	require.NoError(t, err2)
-	input := keeper.CreateTestEnv(t)
+	input := testutil.CreateTestEnv(t)
 	sdkCtx := input.Context
 	ctx := sdk.WrapSDKContext(input.Context)
 	k := input.QgbKeeper
@@ -94,7 +94,7 @@ func TestAllValsetConfirmsByNonce(t *testing.T) {
 		myValidatorEthereumAddr3, _ = stakingtypes.NewEthAddress("0x0303030303030303030303030303030303030303")
 	)
 
-	input := keeper.CreateTestEnv(t)
+	input := testutil.CreateTestEnv(t)
 	sdkCtx := input.Context
 	ctx := sdk.WrapSDKContext(input.Context)
 	k := input.QgbKeeper
@@ -153,27 +153,27 @@ func TestQueryCurrentValset(t *testing.T) {
 		Members: []types.BridgeValidator{
 			{
 				Power:           858993459,
-				EthereumAddress: keeper.EthAddrs[0].GetAddress(),
+				EthereumAddress: testutil.EthAddrs[0].GetAddress(),
 			},
 			{
 				Power:           858993459,
-				EthereumAddress: keeper.EthAddrs[1].GetAddress(),
+				EthereumAddress: testutil.EthAddrs[1].GetAddress(),
 			},
 			{
 				Power:           858993459,
-				EthereumAddress: keeper.EthAddrs[2].GetAddress(),
+				EthereumAddress: testutil.EthAddrs[2].GetAddress(),
 			},
 			{
 				Power:           858993459,
-				EthereumAddress: keeper.EthAddrs[3].GetAddress(),
+				EthereumAddress: testutil.EthAddrs[3].GetAddress(),
 			},
 			{
 				Power:           858993459,
-				EthereumAddress: keeper.EthAddrs[4].GetAddress(),
+				EthereumAddress: testutil.EthAddrs[4].GetAddress(),
 			},
 		},
 	}
-	input, _ := keeper.SetupFiveValChain(t)
+	input, _ := testutil.SetupFiveValChain(t)
 	sdkCtx := input.Context
 
 	currentValset, err := input.QgbKeeper.GetCurrentValset(sdkCtx)
